Improve sign-compare fixes: avoid casting size() to int
Replace the previous (int)size() casts with proper C++ idioms:

- Range-based for loops where the index is only used to access elements:
  cGame_logic, cMap (init + draw_bullets), cDrawManager, cIniGameRessouces

- size_t loop variable with static_cast<int> at the point of use where
  the index must be stored or returned as int:
  cParticleDrawer (push_back into vector<int>),
  cParticle::findNewSlot (returns int),
  createBullet in common.cpp (assigns to int new_id)

- Cast the signed value rather than size() for genuine mixed comparisons:
  cParticle::create: iType >= 0 && static_cast<size_t>(iType) < size()
  ini.cpp: id >= 0 && static_cast<size_t>(id) >= size()

+++ src/utils/common.cpp
@@ -185,9 +185,9 @@ int createBullet(int type, int fromCell, int targetCell, int unitWhichShoots, in
 {
     int new_id = -1;
 
-    for (int i = 0; i < (int)game.g_Bullets.size(); i++)
+    for (size_t i = 0; i < game.g_Bullets.size(); i++)
         if (game.g_Bullets[i].bAlive == false) {
-            new_id = i;
+            new_id = static_cast<int>(i);
             break;
         }
